#38/refactoring: 중복 클래스 정리 및 restapi 통일
