+++ src/main/java/com/sesac2ndproject/attendancemanagementsystem/domain/course/repository/EnrollmentRepository.java
@@ -1,7 +1,6 @@
 package com.sesac2ndproject.attendancemanagementsystem.domain.course.repository;
 
-import com.sesac2ndproject.attendancemanagementsystem.domain.admin.dto.ResponseAttendanceByDateDTO;
-import com.sesac2ndproject.attendancemanagementsystem.domain.admin.dto.ResponseByDateAndCourseIdDTO;
+import com.sesac2ndproject.attendancemanagementsystem.domain.attendance.query.dto.ResponseAttendanceFlatDTO;
 import com.sesac2ndproject.attendancemanagementsystem.domain.course.entity.Enrollment;
 import com.sesac2ndproject.attendancemanagementsystem.global.type.EnrollmentStatus;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -13,37 +12,59 @@ import java.util.List;
 
 public interface EnrollmentRepository extends JpaRepository<Enrollment, Long> {
 
-
-    // Team D 쿼리 1 : 해당하는 courseId를 가지고 Member들만 찾아오기.
+    // Team D 쿼리 1 : 해당하는 courseId를 가진 Enrollment(Member 포함) 찾아오기
     List<Enrollment> findMemberIdByCourseId(Long courseId);
 
-    // Team D 쿼리 2 : *통합 출석부 조회:** 날짜 + 과정ID를 받으면 → 해당 수강생들의 `DailyAttendance`와 `DetailedAttendance`를 조인(또는 Fetch)하여 가져오기.
-    @Query("SELECT new com.sesac2ndproject.attendancemanagementsystem.domain.admin.dto.ResponseByDateAndCourseIdDTO(" +
-            " da.id, da.memberId, mem.name, enr.course.id, cor.courseName, da.date, da.status, dea)" +
+    /**
+     * Team D 쿼리 2 : 통합 출석부 조회 (날짜 + 과정ID)
+     * [수정됨] ResponseByDateAndCourseIdDTO 삭제 -> ResponseAttendanceByDateDTO.FlatResponse 사용
+     * 참고: JPQL에서 Inner Class 생성자를 호출할 때는 '$' 기호를 사용합니다.
+     */
+    @Query("SELECT new com.sesac2ndproject.attendancemanagementsystem.domain.attendance.query.dto.ResponseAttendanceFlatDTO(" +
+            "   da.id, " +
+            "   da.memberId, " +
+            "   mem.name, " +
+            "   enr.course.id, " +
+            "   cor.courseName, " +
+            "   da.date, " +
+            "   da.status, " +
+            "   dea" +
+            ") " +
             " FROM DailyAttendance AS da" +
             " JOIN Enrollment AS enr ON da.memberId = enr.member.id" +
-            " LEFT JOIN DetailedAttendance AS dea ON da.id = dea.dailyAttendanceId" +  /* 상세 기록이 없어도 가져오도록 LEFT JOIN으로 처리. 단 NULL값 주의 */
             " JOIN Member mem ON da.memberId = mem.id" +
             " JOIN Course cor ON enr.course.id = cor.id" +
+            " LEFT JOIN DetailedAttendance AS dea ON da.id = dea.dailyAttendanceId" +
             " WHERE da.date = :workDate AND enr.course.id = :courseId")
-    List<ResponseByDateAndCourseIdDTO> integratedAttendance(@Param("workDate") LocalDate workDate, @Param("courseId") Long courseId);
+    List<ResponseAttendanceFlatDTO> integratedAttendance(@Param("workDate") LocalDate workDate, @Param("courseId") Long courseId);
 
-    // Team D 쿼리 3 :
-    @Query("SELECT new com.sesac2ndproject.attendancemanagementsystem.domain.admin.dto.ResponseAttendanceByDateDTO$FlatResponse(" +
-            "da.id, da.memberId, mem.name, cor.id, cor.courseName, da.date, da.status, dea) " +
+    /**
+     * Team D 쿼리 3 : (쿼리 2와 기능 중복, 로직 통합 권장)
+     * 기존에 잘 작성되어 있던 FlatResponse 반환 메서드입니다.
+     */
+    @Query("SELECT new com.sesac2ndproject.attendancemanagementsystem.domain.attendance.query.dto.ResponseAttendanceFlatDTO(" +
+            "   da.id, " +
+            "   da.memberId, " +
+            "   mem.name, " +
+            "   cor.id, " +
+            "   cor.courseName, " +
+            "   da.date, " +
+            "   da.status, " +
+            "   dea" +
+            ") " +
             " FROM DailyAttendance AS da" +
-            // Daily(Long) <=> Enrollment(Member객체) 연결: er.member.memberId 사용
             " JOIN Enrollment AS er ON da.memberId = er.member.id" +
             " JOIN Member AS mem ON da.memberId = mem.id" +
             " JOIN Course AS cor ON da.courseId = cor.id" +
-            // Daily <=> Detailed 연결: attendanceId 사용 (LEFT JOIN 권장: 상세기록 없어도 출석부는 나오게)
             " LEFT JOIN DetailedAttendance AS dea ON da.id = dea.dailyAttendanceId" +
-            // Enrollment(Course객체) <=> 파라미터(Long) 비교: er.course.courseId 사용
             " WHERE da.date = :workDate AND er.course.id = :courseId")
-    List<ResponseAttendanceByDateDTO.FlatResponse> findIntegratedAttendanceFlat(@Param("workDate") LocalDate workDate, @Param("courseId") Long courseId);
+    List<ResponseAttendanceFlatDTO> findIntegratedAttendanceFlat(@Param("workDate") LocalDate workDate, @Param("courseId") Long courseId);
 
-    // Team D 쿼리 4 : CSV, Excel 파일 변환을 위한 조건 없는 전체 통합 출석부 조회(WHERE절 없음, 날짜 최신순으로 정렬됨)
-    @Query("SELECT new com.sesac2ndproject.attendancemanagementsystem.domain.admin.dto.ResponseByDateAndCourseIdDTO(" +
+    /**
+     * Team D 쿼리 4 : 전체 통합 출석부 조회 (다운로드용)
+     * [수정됨] ResponseByDateAndCourseIdDTO 삭제 -> ResponseAttendanceByDateDTO.FlatResponse 사용
+     */
+    @Query("SELECT new com.sesac2ndproject.attendancemanagementsystem.domain.attendance.query.dto.ResponseAttendanceFlatDTO(" +
             "   da.id, " +
             "   da.memberId, " +
             "   mem.name, " +
@@ -55,12 +76,13 @@ public interface EnrollmentRepository extends JpaRepository<Enrollment, Long> {
             ") " +
             " FROM DailyAttendance da" +
             " JOIN Enrollment enr ON da.memberId = enr.member.id" +
-            " LEFT JOIN DetailedAttendance dea ON da.id = dea.dailyAttendanceId" +
             " JOIN Member mem ON da.memberId = mem.id" +
             " JOIN Course cor ON enr.course.id = cor.id" +
+            " LEFT JOIN DetailedAttendance dea ON da.id = dea.dailyAttendanceId" +
             " ORDER BY da.date DESC, da.memberId ASC")
-    List<ResponseByDateAndCourseIdDTO> findAllIntegratedAttendance();
+    List<ResponseAttendanceFlatDTO> findAllIntegratedAttendance();
 
+    // 특정 상태의 수강생 ID 목록 조회
     @Query("SELECT e.member.id FROM Enrollment e WHERE e.course.id = :courseId AND e.status = :status")
     List<Long> findMemberIdsByCourseIdAndStatus(
             @Param("courseId") Long courseId,
