+++ src/main/java/com/sesac2ndproject/attendancemanagementsystem/global/type/AttendanceStatus.java
@@ -10,7 +10,8 @@ public enum AttendanceStatus {
     LATE("지각"),
     ABSENT("결석"),
     NONE("미체크"),  // 아직 시간이 안 돼서 안 찍은 상태
-    LEAVE("조퇴");
+    LEAVE("조퇴"),
+    OFFICIAL_LEAVE("공결");
 
     private final String description;
 }
